requirements.txt
streamlit==1.32.0
pandas==2.2.1
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
openai
beautifulsoup4==4.12.3
python-dateutil==2.8.2
langchain-openai
langchain
langgraph>=0.0.5
watchfiles>=0.21.0
ipython>=8.0.0
elasticsearch==8.13.0
msal
msgraph-core==0.2.2
requests